+++ Face.h
@@ -1,3 +1,16 @@
+///////////////////////////////////////////////////
+//
+//
+//  Created by Jake Meegan on 16/10/2019.
+//
+//	------------------------
+//	Face.h
+//	------------------------
+//
+//	A minimal class for storing a face with its associated vertex ID's
+//
+///////////////////////////////////////////////////
+
 #ifndef FACE_H
 #define FACE_H
 
@@ -10,8 +23,9 @@ class Face
 	// the coordinates
     unsigned long v1, v2, v3;
 
-	void setVertex(unsigned int vertex, unsigned long value);
-	unsigned long & getVertex(unsigned int vertex);
+    // Get/Set vertex via ID (0 -> v1, 1 -> v2, 2 -> v3)
+	void setVertex(unsigned long vertex, unsigned long value);
+	unsigned long & getVertex(unsigned long vertex);
 
 	}; // Face
